New SimDrive that extends Drive
